fix: remove `cli-` prefix from CLI package versions
SwiftPM doesn't have versioning other than git tags, so it needs to be a
"clean" semantic version. For the web package we are explicitly
publishing to hex.pm and have a version inside the MixProject that's
separate from the git tag, so we are keeping the `web-` prefix there to
avoid collision.

Signed-off-by: Christoph Schmatzler <christoph@tuist.dev>